экспорты и импорты из денкирса, немного поправить и заработает

+++ src/main/java/com/excelconverterweb/repository/IlluminatorRepository.java
@@ -3,6 +3,5 @@ package com.excelconverterweb.repository;
 import com.excelconverterweb.model.Illuminator;
 import org.springframework.data.repository.CrudRepository;
 
-
-public interface IlluminatorRepository extends CrudRepository<Illuminator, Long> {
+public interface IlluminatorRepository extends CrudRepository<Illuminator, Integer> {
 }
